+++ src/lldb/Bindings/SBCommunicationBinding.h
@@ -43,10 +43,10 @@ LLDB_API bool SBCommunicationGetCloseOnEOF(SBCommunicationRef instance);
 
 LLDB_API void SBCommunicationSetCloseOnEOF(SBCommunicationRef instance, bool b);
 
-LLDB_API unsigned int SBCommunicationRead(SBCommunicationRef instance, void *dst, size_t dst_len,
+LLDB_API size_t SBCommunicationRead(SBCommunicationRef instance, void *dst, size_t dst_len,
                                           uint32_t timeout_usec, ENUM(ConnectionStatus) status);
 
-LLDB_API unsigned int SBCommunicationWrite(SBCommunicationRef instance, void *src, size_t src_len,
+LLDB_API size_t SBCommunicationWrite(SBCommunicationRef instance, void *src, size_t src_len,
                                            ENUM(ConnectionStatus) status);
 
 LLDB_API bool SBCommunicationReadThreadStart(SBCommunicationRef instance);
