+++ src/lldb/Bindings/SBSectionBinding.h
@@ -30,7 +30,7 @@ LLDB_API SBSectionRef SBSectionGetParent(SBSectionRef instance);
 
 LLDB_API SBSectionRef SBSectionFindSubSection(SBSectionRef instance, const char *sect_name);
 
-LLDB_API unsigned int SBSectionGetNumSubSections(SBSectionRef instance);
+LLDB_API size_t SBSectionGetNumSubSections(SBSectionRef instance);
 
 LLDB_API SBSectionRef SBSectionGetSubSectionAtIndex(SBSectionRef instance, size_t idx);
 
@@ -40,9 +40,9 @@ LLDB_API lldb_addr_t SBSectionGetLoadAddress(SBSectionRef instance, SBTargetRef 
 
 LLDB_API lldb_addr_t SBSectionGetByteSize(SBSectionRef instance);
 
-LLDB_API unsigned long long SBSectionGetFileOffset(SBSectionRef instance);
+LLDB_API uint64_t SBSectionGetFileOffset(SBSectionRef instance);
 
-LLDB_API unsigned long long SBSectionGetFileByteSize(SBSectionRef instance);
+LLDB_API uint64_t SBSectionGetFileByteSize(SBSectionRef instance);
 
 LLDB_API SBDataRef SBSectionGetSectionData(SBSectionRef instance);
 
@@ -50,7 +50,7 @@ LLDB_API SBDataRef SBSectionGetSectionData2(SBSectionRef instance, uint64_t offs
 
 LLDB_API ENUM(SectionType) SBSectionGetSectionType(SBSectionRef instance);
 
-LLDB_API unsigned int SBSectionGetTargetByteSize(SBSectionRef instance);
+LLDB_API uint32_t SBSectionGetTargetByteSize(SBSectionRef instance);
 
 LLDB_API bool SBSectionGetDescription(SBSectionRef instance, SBStreamRef description);
